pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.40.0",
    "face-recognition>=1.3.0",
    "flask>=3.1.3",
    "flask-socketio>=5.5.1",
    "numpy>=2.4.4",
    "opencv-python>=4.13.0.92",
    "pillow>=12.2.0",
    "setuptools<81",
    "watchdog>=6.0.0",
    "werkzeug>=3.1.8",
]
